frontend/requirements.txt
# Streamlit Framework
streamlit>=1.30.0

# API Client
requests>=2.31.0
httpx>=0.26.0

# Data Processing
pandas>=2.1.4
openpyxl==3.1.2

# PDF Generation
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
